[#163] Move DescriptorFactory into a new package - context.
