+++ src/main/java/cz/cvut/kbss/termit/persistence/dao/AssetDao.java
@@ -21,7 +21,7 @@ import cz.cvut.kbss.termit.exception.PersistenceException;
 import cz.cvut.kbss.termit.model.Asset;
 import cz.cvut.kbss.termit.model.User;
 import cz.cvut.kbss.termit.model.comment.Comment;
-import cz.cvut.kbss.termit.persistence.DescriptorFactory;
+import cz.cvut.kbss.termit.persistence.context.DescriptorFactory;
 import cz.cvut.kbss.termit.util.Vocabulary;
 import cz.cvut.kbss.termit.util.Configuration.Persistence;
