+++ src/test/java/cz/cvut/kbss/termit/persistence/dao/BaseTermDaoTestRunner.java
@@ -3,8 +3,7 @@ package cz.cvut.kbss.termit.persistence.dao;
 import cz.cvut.kbss.jopa.model.EntityManager;
 import cz.cvut.kbss.termit.environment.Generator;
 import cz.cvut.kbss.termit.model.Vocabulary;
-import cz.cvut.kbss.termit.persistence.DescriptorFactory;
-import org.junit.jupiter.api.BeforeEach;
+import cz.cvut.kbss.termit.persistence.context.DescriptorFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public class BaseTermDaoTestRunner extends BaseDaoTestRunner {
